==> t5-question-answering/1_PreprocessDataset.ipynb <==
!pip install --upgrade pip
!pip install 'farm-haystack'
!pip install datasets==1.0.2
!pip install tqdm==4.57.0

import pandas as pd
pd.set_option("display.max_colwidth", -1)

# connect your personal google drive to store dataset and trained model
from google.colab import drive
drive.mount('/content/gdrive/')

**Preprocessing the Biology textbook Chapter 2 - Chemistry of life text**

#--Using Haystack API to 
from haystack import Pipeline
from haystack.nodes import PreProcessor,TextConverter

processor = PreProcessor(
    clean_empty_lines=True,
    clean_whitespace=True,
    clean_header_footer=True,
    split_by="word",
    split_length=100,
    split_respect_sentence_boundary=True,
    split_overlap=5
)

converter = TextConverter();
doc = converter.convert(file_path='/content/gdrive/My Drive/CS 677 Project/dataset/chemistryoflife_v3.txt', meta=None)
docs = processor.process(doc)

df = pd.DataFrame(docs)
df.shape

df.head()
df.to_excel("/content/gdrive/My Drive/CS 677 Project/dataset/chemistryoflife_all_v3.xlsx",
             sheet_name='Sheet_name_1') 

chem = df['content']

chem.shape

chem.head()

chem.to_excel("/content/gdrive/My Drive/CS 677 Project/dataset/chemistryoflife_v3.xlsx",
             sheet_name='Sheet_name_1')  

**Preparing the SQUAD dataset for validation and comparison**

from datasets import load_dataset
train_dataset = load_dataset('squad', split='train')
valid_dataset = load_dataset('squad', split='validation')

from pprint import pprint 
sample_validation_dataset = next(iter(valid_dataset))
pprint (sample_validation_dataset)

context = sample_validation_dataset['context']
question = sample_validation_dataset['question']
answer = sample_validation_dataset['answers']['text'][0]

print ("context: ",context)
print ("question: ",question)
print ("answer: ",answer)

df_train = pd.DataFrame( columns = ['context', 'answer','question'])
df_validation = pd.DataFrame( columns = ['context', 'answer','question'])
print (df_validation)
print (df_train)

from tqdm.notebook import tqdm
count_long = 0
count_short = 0

        
for index,val in enumerate(tqdm(train_dataset)):
    passage = val['context']
    question = val['question']
    answer = val['answers']['text'][0]
    no_of_words = len(answer.split())
    if no_of_words >= 7:
        count_long = count_long + 1
        continue
    else:
        df_train.loc[count_short]= [passage] + [answer] + [question] 
        count_short = count_short + 1       

print ("count_long train dataset: ",count_long)
print ("count_short train dataset: ",count_short)

count_long = 0
count_short = 0

        
for index,val in enumerate(tqdm(valid_dataset)):
    passage = val['context']
    question = val['question']
    answer = val['answers']['text'][0]
    no_of_words = len(answer.split())
    if no_of_words >= 7:
        count_long = count_long + 1
        continue
    else:
        df_validation.loc[count_short]= [passage] + [answer] + [question] 
        count_short = count_short + 1       

print ("count_long validation dataset: ",count_long)
print ("count_short validation dataset: ",count_short)

from sklearn.utils import shuffle
df_train = shuffle(df_train)
df_validation = shuffle(df_validation)

print (df_train.shape)
print (df_validation.shape)

df_train.head()

train_save_path = '/content/gdrive/My Drive/CS 677 Project/dataset/squad_t5_train.csv'
validation_save_path = '/content/gdrive/My Drive/CS 677 Project/dataset/squad_t5_val.csv'
df_train.to_csv(train_save_path, index = False)
df_validation.to_csv(validation_save_path, index = False)
