==> t5-question-answering/2_ExtractKeywords.ipynb <==
**Keyword Extraction using KeyBert and Yake**

Reference: https://www.analyticsvidhya.com/blog/2022/03/keyword-extraction-methods-from-documents-in-nlp/

!pip3 install keybert
!pip3 install yake

# connect your personal google drive to store dataset and trained model
from google.colab import drive
drive.mount('/content/gdrive/')

import pandas as pd
pd.set_option("display.max_colwidth", -1)
from keybert import KeyBERT
import yake

# Define all functions

def get_source_from_excel(dataset):
  source=pd.read_excel(dataset, sheet_name='Sheet_name_1')
  return source

def keybert_extract(context):
  kw_BERT = KeyBERT()
  bert_keywords = kw_BERT.extract_keywords(context)
  bert_df = pd.DataFrame(bert_keywords, columns =['keyword', 'normalized_weight'])
  return bert_df

def yake_extract(context):
  kw_yake = yake.KeywordExtractor(n=2,top=5)
  yake_keywords = kw_yake.extract_keywords(context)
  yake_df = pd.DataFrame(yake_keywords, columns =['keyword', 'weight']).sort_values(by=['weight', 'keyword'], ascending=[True, False])
  yake_df['normalized_weight'] = 1 - ( yake_df['weight']  / sum(yake_df['weight']))
  yake_df.drop(['weight'], axis=1, inplace = True)
  return yake_df

def get_keywords_from_context(context):
  keywords_df = pd.concat([yake_extract(context), keybert_extract(context)], axis=0)    
  return keywords_df

def get_keywords_for_source(source):
  context_keywords = pd.DataFrame()
  for index, row in source.iterrows():
    context = row['content'] 
    keywords = get_keywords_from_context(context)
    c_keywords  = pd.DataFrame()
    c_keywords['context'] = [context]
    c_keywords = keywords.merge(c_keywords, how='cross')
    context_keywords = pd.concat([context_keywords , c_keywords], axis=0)
  return  context_keywords


#Usage
source = get_source_from_excel("/content/gdrive/My Drive/CS 677 Project/dataset/chemistryoflife_v3.xlsx")
context_keywords = get_keywords_for_source(source)
context_keywords.to_excel("/content/gdrive/My Drive/CS 677 Project/dataset/context_keywords.xlsx", sheet_name='Sheet_name_1')
