==> t5-question-answering/5_CollectValidQAPairs.ipynb <==
**Model 1 **: https://huggingface.co/sentence-transformers/all-MiniLM-L6-v2

**Model 2:** https://huggingface.co/google/flan-t5-xxl

Metric used for Evaluation : Semantic Answer Similarity (SAS)

Semantic Answer Similarity (SAS): https://arxiv.org/abs/2108.06130

pip install -U sentence-transformers

from sentence_transformers import SentenceTransformer, util

#Load the model
model = SentenceTransformer('sentence-transformers/multi-qa-distilbert-dot-v1')

#------------Sample Implementation for Sentence Similarity -----------------

#Initialize the inputs
#query = "Atoms are the building blocks that come together through chemical bonding to form what in the universe?"
#answers = ["Molecule", "dummy"]
#
#query = "How do atoms come together?"
#answers = ["chemical bonding", "when they are bonded together"]
#
#query = "What is the model of atoms in the universe?"
#answers = ["molecule", "Carbon (black), hydrogen (white), nitrogen (blue), oxygen (red) and sulfur" , "atoms are made of protons and neutrons"]

query = "What size are the atoms in the model of a molecule?"
answers = ["atoms are the size of a molecule","atomic","proportional"]

#Encode query and documents
query_emb = model.encode(query)
doc_emb = model.encode(answers)

#Compute dot score between query and all document embeddings
scores = util.dot_score(query_emb, doc_emb)[0].cpu().tolist()

#Combine docs & scores
doc_score_pairs = list(zip(answers, scores))

#Sort by decreasing score
doc_score_pairs = sorted(doc_score_pairs, key=lambda x: x[1], reverse=True)

#Output passages & scores
for doc, score in doc_score_pairs:
    print(score, doc)

# connect your personal google drive to store dataset and trained model
from google.colab import drive
drive.mount('/content/gdrive/')

# Read the excel which contains the context, original keyword, generated question and the predicted answer
import pandas as pd
source=pd.read_excel("/content/gdrive/My Drive/CS 677 Project/dataset/predicted_answer.xlsx", sheet_name='Sheet_name_1')

# Spot check the data
source.head()

# Function to calculate similarity score between the original keyword and the predicted answer

def get_similarity_scores(question, answers):
  #Encode query and documents
  query_emb = model.encode(question)
  answer_emb = model.encode(answers)
  #Compute dot score between query and all document embeddings
  scores = util.dot_score(query_emb, answer_emb)[0].cpu().tolist()
  return scores

# Function to compare two similarity scores based on a threshold
threshold = 10

def check_similarity(score1, score2):
  if abs(score1-score2) <=threshold :
    return 'Match'
  else:
    return 'No Match'

# Spot check the function using an example
similarity_scores = get_similarity_scores('Atoms are the building blocks that come together through chemical bonding to do what in the universe?', 
                 ['form molecules','ten form molecules'])

check_similarity(similarity_scores[0], similarity_scores[1])

# Collect Valid QA pairs

QA_pairs = []
for index, row in source.iterrows():
  context = row[0] 
  original_keyword = row[1]
  question = row[2]
  predicted_answer = row[3]
  answers = [original_keyword, predicted_answer]
  similarity_scores = get_similarity_scores(question, answers)
  check = check_similarity(similarity_scores[0], similarity_scores[1])
  if check == 'Match':
    valid_QA_tuple = (context, question, predicted_answer, similarity_scores[1] )
    QA_pairs.append(valid_QA_tuple)

# Spot check valid QA pair
QA_pairs[:2]

# Select only distinct QA pairs
dist_QA_pairs = list(set(QA_pairs))

# Spot check distinct QA pair
dist_QA_pairs[:2]

# Write the final output to excel

final_QA_pairs = pd.DataFrame(dist_QA_pairs)
final_QA_pairs.to_excel("/content/gdrive/My Drive/CS 677 Project/dataset/final_QA_pairs.xlsx", sheet_name='Sheet_name_1')
